refactor: rename `model` package to `domain`
