+++ src/main/java/it/pagopa/transactions/data/RptIdWritingStringConverter.java
@@ -1,6 +1,6 @@
 package it.pagopa.transactions.data;
 
-import it.pagopa.transactions.model.RptId;
+import it.pagopa.transactions.domain.RptId;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.data.convert.WritingConverter;
 import org.springframework.stereotype.Component;
