+++ src/main/java/it/pagopa/transactions/domain/TransactionAmount.java
@@ -1,4 +1,4 @@
-package it.pagopa.transactions.model;
+package it.pagopa.transactions.domain;
 
 import it.pagopa.transactions.annotations.ValueObject;
