+++ src/main/java/it/pagopa/transactions/repositories/TransactionTokensRepository.java
@@ -1,6 +1,6 @@
 package it.pagopa.transactions.repositories;
 
-import it.pagopa.transactions.model.RptId;
+import it.pagopa.transactions.domain.RptId;
 import org.springframework.data.repository.CrudRepository;
 
 public interface TransactionTokensRepository extends CrudRepository<TransactionTokens, RptId> {}
