+++ src/test/java/it/pagopa/transactions/commands/handlers/TransactionInitializerHandlerTest.java
@@ -11,8 +11,8 @@ import it.pagopa.transactions.client.EcommerceSessionsClient;
 import it.pagopa.transactions.client.NodeForPspClient;
 import it.pagopa.transactions.commands.TransactionsCommand;
 import it.pagopa.transactions.documents.TransactionInitData;
-import it.pagopa.transactions.model.IdempotencyKey;
-import it.pagopa.transactions.model.RptId;
+import it.pagopa.transactions.domain.IdempotencyKey;
+import it.pagopa.transactions.domain.RptId;
 import it.pagopa.transactions.projections.TransactionsProjection;
 import it.pagopa.transactions.repositories.TransactionTokens;
 import it.pagopa.transactions.repositories.TransactionTokensRepository;
